feat: smarter balance assertion dates
We find the statement's end date from the OFX file. However, banks and
credit cards typically have pending transactions that are not included
in downloads. When we download the next statement, new transactions may
appear prior to the balance assertion date that we generate for this
statement. To attempt to avoid this, we set the balance assertion date
to either two days before the statement's end date or the last
transaction's date, whichever is later.